pyproject.toml
[tool.poetry]
name = "python-pdf-text-extractor"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pdfminer-six = "^20221105"
celery = {extras = ["redis"], version = "^5.3.4"}
uvicorn = "^0.23.2"
fastapi = "^0.104.0"
python-multipart = "^0.0.6"
python-magic = "^0.4.27"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportUnknownVariableType = false
reportUnknownMemberType = false
reportMissingTypeStubs = false
reportUnknownArgumentType = false
